requirements.txt
streamlit==1.32.2
feedparser
requests
beautifulsoup4
pandas
yfinance
# ─ 본문 파싱용 ─
trafilatura==2.0.0
readability-lxml
lxml[html_clean]>=5.4
newspaper3k
# ─ 요약 모델 ─
transformers
sentencepiece
torch
# ─ 기타 ─
streamlit-autorefresh
nltk
selenium